WIP api books
